pyproject.toml
[project]
name = "core-engine"
version = "0.1.0"
description = "Core Engine MVP - Claude Agent SDK wrapper with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.0.0",
    "filelock>=3.16.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "claude-code-sdk>=0.0.25",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
